NAS_Migration_PoC/requirements.txt
# Core Extraction
PyMuPDF==1.25.1
pytesseract==0.3.10
Pillow==11.0.0
openpyxl==3.1.5
python-docx==1.1.2

# Database (PostgreSQL for Prod, SQLite built-in for PoC)
psycopg2-binary==2.9.10

# AI/LLM Integration
ollama==0.4.5
openai==1.59.3

# Utilities
tqdm==4.67.1
python-dateutil==2.9.0.post0
pyyaml==6.0.2
pandas==2.2.3

# Backend (FastAPI + Realtime)
fastapi==0.111.0
uvicorn==0.30.1
watchdog==4.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
